Rename action files
