+++ _common/gameModel/libs/ClientScript/dashboard/dashboardFacade.ts
@@ -59,7 +59,7 @@ import { dashboardLogger } from '../tools/logger';
 import { MapEntityActivable } from '../game/common/simulationState/activableState';
 import { getActionTemplates } from '../game/mainSimulationLogic';
 import { MapChoiceActionTemplate } from '../game/common/actions/actionTemplate/actionTemplateMap';
-import { CasuMessageAction } from '../game/common/actions/actionRadio';
+import { CasuMessageAction } from '../game/common/actions/radioActions';
 
 // -------------------------------------------------------------------------------------------------
 // state part
